+++ format.c
@@ -248,7 +248,7 @@ HB_FUNC( FORMAT_SET_ITALIC )
 HB_FUNC( FORMAT_SET_UNDERLINE )
 { 
    lxw_format *self = hb_parptr( 1 ) ;
-   uint8_t style = hb_parni( 2 ) ;
+   uint8_t style = (uint8_t) hb_parni( 2 ) ;
 
    format_set_underline( self, style ); 
 }
@@ -283,7 +283,7 @@ HB_FUNC( FORMAT_SET_FONT_STRIKEOUT )
 HB_FUNC( FORMAT_SET_FONT_SCRIPT )
 { 
    lxw_format *self = hb_parptr( 1 ) ;
-   uint8_t style = hb_parni( 2 ) ;
+   uint8_t style = (uint8_t) hb_parni( 2 ) ;
 
    format_set_font_script( self, style ); 
 }
@@ -387,7 +387,7 @@ HB_FUNC( FORMAT_SET_HIDDEN )
 HB_FUNC( FORMAT_SET_ALIGN )
 { 
    lxw_format *self = hb_parptr( 1 ) ;
-   uint8_t value = hb_parni( 2 ) ;
+   uint8_t value = (uint8_t) hb_parni( 2 ) ;
 
    format_set_align( self, value ); 
 }
@@ -422,7 +422,7 @@ HB_FUNC( FORMAT_SET_TEXT_WRAP )
 HB_FUNC( FORMAT_SET_ROTATION )
 { 
    lxw_format *self = hb_parptr( 1 ) ;
-   int16_t angle = hb_parnl( 2 ) ;
+   int16_t angle = (int16_t) hb_parnl( 2 ) ;
 
    format_set_rotation( self, angle ); 
 }
@@ -440,7 +440,7 @@ HB_FUNC( FORMAT_SET_ROTATION )
 HB_FUNC( FORMAT_SET_INDENT )
 { 
    lxw_format *self = hb_parptr( 1 ) ;
-   uint8_t value = hb_parni( 2 ) ;
+   uint8_t value = (uint8_t) hb_parni( 2 ) ;
 
    format_set_indent( self, value ); 
 }
@@ -493,7 +493,7 @@ HB_FUNC( FORMAT_SET_TEXT_JUSTLAST )
 HB_FUNC( FORMAT_SET_PATTERN )
 { 
    lxw_format *self = hb_parptr( 1 ) ;
-   uint8_t value = hb_parni( 2 ) ;
+   uint8_t value = (uint8_t) hb_parni( 2 ) ;
 
    format_set_pattern( self, value ); 
 }
@@ -547,7 +547,7 @@ HB_FUNC( FORMAT_SET_FG_COLOR )
 HB_FUNC( FORMAT_SET_BORDER )
 { 
    lxw_format *self = hb_parptr(1 ) ;
-   uint8_t style = hb_parni( 2 ) ;
+   uint8_t style = (uint8_t) hb_parni( 2 ) ;
 
    format_set_border( self, style ); 
 }
@@ -583,7 +583,7 @@ HB_FUNC( FORMAT_SET_BORDER_COLOR )
 HB_FUNC( FORMAT_SET_BOTTOM )
 { 
    lxw_format *self = hb_parptr( 1 ) ;
-   uint8_t style = hb_parni( 2 ) ;
+   uint8_t style = (uint8_t) hb_parni( 2 ) ;
 
    format_set_bottom( self, style ); 
 }
@@ -619,7 +619,7 @@ HB_FUNC( FORMAT_SET_BOTTOM_COLOR )
 HB_FUNC( FORMAT_SET_LEFT )
 { 
    lxw_format *self = hb_parptr( 1 ) ;
-   uint8_t style = hb_parni( 2 ) ;
+   uint8_t style = (uint8_t) hb_parni( 2 ) ;
 
    format_set_left( self, style ); 
 }
@@ -655,7 +655,7 @@ HB_FUNC( FORMAT_SET_LEFT_COLOR )
 HB_FUNC( FORMAT_SET_RIGHT )
 { 
    lxw_format *self = hb_parptr(1 ) ;
-   uint8_t style = hb_parni( 2 ) ;
+   uint8_t style = (uint8_t) hb_parni( 2 ) ;
 
    format_set_right( self, style ); 
 }
@@ -691,7 +691,7 @@ HB_FUNC( FORMAT_SET_RIGHT_COLOR )
 HB_FUNC( FORMAT_SET_TOP )
 { 
    lxw_format *self = hb_parptr( 1 ) ;
-   uint8_t style = hb_parni( 2 ) ;
+   uint8_t style = (uint8_t) hb_parni( 2 ) ;
 
    format_set_top( self, style ); 
 }
@@ -727,7 +727,7 @@ HB_FUNC( FORMAT_SET_TOP_COLOR )
 HB_FUNC( FORMAT_SET_DIAG_TYPE )
 { 
    lxw_format *self = hb_parptr( 1 ) ;
-   uint8_t type = hb_parni( 2 ) ;
+   uint8_t type = (uint8_t) hb_parni( 2 ) ;
 
    format_set_diag_type( self, type ); 
 }
@@ -763,7 +763,7 @@ HB_FUNC( FORMAT_SET_DIAG_COLOR )
 HB_FUNC( FORMAT_SET_DIAG_BORDER )
 { 
    lxw_format *self = hb_parptr( 1 ) ;
-   uint8_t style = hb_parni( 2 ) ;
+   uint8_t style = (uint8_t) hb_parni( 2 ) ;
 
    format_set_diag_border( self, style ); 
 }
@@ -781,7 +781,7 @@ HB_FUNC( FORMAT_SET_DIAG_BORDER )
 HB_FUNC( FORMAT_SET_NUM_FORMAT_INDEX )
 { 
    lxw_format *self = hb_parptr( 1 ) ;
-   uint8_t value = hb_parni( 2 ) ;
+   uint8_t value = (uint8_t) hb_parni( 2 ) ;
 
    format_set_num_format_index( self, value ); 
 }
@@ -800,7 +800,7 @@ HB_FUNC( FORMAT_SET_NUM_FORMAT_INDEX )
 HB_FUNC( FORMAT_SET_VALIGN )
 { 
    lxw_format *self = hb_parptr( 1 ) ;
-   uint8_t value = hb_parni( 2 ) ;
+   uint8_t value = (uint8_t) hb_parni( 2 ) ;
 
    format_set_valign( self, value ); 
 }
@@ -818,7 +818,7 @@ HB_FUNC( FORMAT_SET_VALIGN )
 HB_FUNC( FORMAT_SET_READING_ORDER )
 { 
    lxw_format *self = hb_parptr( 1 ) ;
-   uint8_t value = hb_parni( 2 ) ;
+   uint8_t value = (uint8_t) hb_parni( 2 ) ;
 
    format_set_reading_order( self, value ); 
 }
@@ -836,7 +836,7 @@ HB_FUNC( FORMAT_SET_READING_ORDER )
 HB_FUNC( FORMAT_SET_FONT_FAMILY )
 { 
    lxw_format *self = hb_parptr( 1 ) ;
-   uint8_t value = hb_parni( 2 ) ;
+   uint8_t value = (uint8_t) hb_parni( 2 ) ;
 
    format_set_font_family( self, value ); 
 }
@@ -854,7 +854,7 @@ HB_FUNC( FORMAT_SET_FONT_FAMILY )
 HB_FUNC( FORMAT_SET_FONT_CHARSET )
 { 
    lxw_format *self = hb_parptr( 1 ) ;
-   uint8_t value = hb_parni( 2 ) ;
+   uint8_t value = (uint8_t) hb_parni( 2 ) ;
 
    format_set_font_charset( self, value ); 
 }
@@ -913,3 +913,4 @@ HB_FUNC( FORMAT_SET_FONT_EXTEND )
 
 
 //eof
+
